+++ Controllers/KullaniciController.cs
@@ -3,12 +3,13 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Muhasebe.Models.Entities;
 using Muhasebe.Services.Interfaces;
 using Microsoft.EntityFrameworkCore;
+using Muhasebe.Common.Helpers;
 
 namespace Muhasebe.Controllers
 {
     public class KullaniciController : Controller
     {
-
+        private const string EntityName = "Kullanıcı";
         private readonly IKullaniciService _kullaniciService;
         private readonly IDepartmanService _departmanService;
         private readonly MuhasebeContext _context;
@@ -20,8 +21,6 @@ namespace Muhasebe.Controllers
             _context = context;
         }
 
-
-
         public async Task<IActionResult> Index()
         {
             var kullanicilar = await _kullaniciService.GetAllKullaniciAsync();
@@ -30,12 +29,7 @@ namespace Muhasebe.Controllers
 
         public async Task<IActionResult> Create()
         {
-            var departmanlar = await _departmanService.GetAllDepartmanAsync();
-            ViewBag.DepartmanId = new SelectList(departmanlar, "DepartmanId", "DepartmanAdi");
-
-            var roller = await _context.Rols.ToListAsync();
-            ViewBag.RolListesi = new SelectList(roller, "RolId", "RolAdi");
-
+            await PopulateSelectListsAsync();
             return View();
         }
 
@@ -45,15 +39,10 @@ namespace Muhasebe.Controllers
             if (ModelState.IsValid)
             {
                 await _kullaniciService.CreateKullaniciAsync(model);
-                return RedirectToAction("Index");
+                return RedirectToAction(nameof(Index));
             }
 
-            var departmanlar = await _departmanService.GetAllDepartmanAsync();
-            ViewBag.DepartmanId = new SelectList(departmanlar, "DepartmanId", "DepartmanAdi", model.DepartmanId);
-
-            var roller = await _context.Rols.ToListAsync();
-            ViewBag.RolListesi = new SelectList(roller, "RolId", "RolAdi", model.RolId);
-
+            await PopulateSelectListsAsync(model.DepartmanId, model.RolId);
             return View(model);
         }
 
@@ -62,27 +51,33 @@ namespace Muhasebe.Controllers
             var kullanici = await _kullaniciService.GetKullaniciByIdAsync(id);
             if (kullanici == null)
             {
-                return NotFound();
+                return this.RecordNotFound(EntityName, "Kullanici", requestedId: id, listLabel: "Kullanıcı listesine dön");
             }
-            var departmanlar = await _departmanService.GetAllDepartmanAsync();
-            ViewBag.DepartmanId = new SelectList(departmanlar, "DepartmanId", "DepartmanAdi", kullanici.DepartmanId);
-            var roller = await _context.Rols.ToListAsync();
-            ViewBag.RolListesi = new SelectList(roller, "RolId", "RolAdi", kullanici.RolId);
+            await PopulateSelectListsAsync(kullanici.DepartmanId, kullanici.RolId);
             return View(kullanici);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Edit(Kullanici model)
+        public async Task<IActionResult> Edit(int id, Kullanici model)
         {
+            if (id != model.KullaniciId)
+            {
+                return this.RecordNotFound(EntityName, "Kullanici", requestedId: id, listLabel: "Kullanıcı listesine dön");
+            }
+
+            var existing = await _kullaniciService.GetKullaniciByIdAsync(id);
+            if (existing == null)
+            {
+                return this.RecordNotFound(EntityName, "Kullanici", requestedId: id, listLabel: "Kullanıcı listesine dön");
+            }
+
             if (ModelState.IsValid)
             {
                 await _kullaniciService.UpdateKullaniciAsync(model);
-                return RedirectToAction("Index");
+                return RedirectToAction(nameof(Index));
             }
-            var departmanlar = await _departmanService.GetAllDepartmanAsync();
-            ViewBag.DepartmanId = new SelectList(departmanlar, "DepartmanId", "DepartmanAdi", model.DepartmanId);
-            var roller = await _context.Rols.ToListAsync();
-            ViewBag.RolListesi = new SelectList(roller, "RolId", "RolAdi", model.RolId);
+
+            await PopulateSelectListsAsync(model.DepartmanId, model.RolId);
             return View(model);
         }
 
@@ -91,7 +86,7 @@ namespace Muhasebe.Controllers
             var kullanici = await _kullaniciService.GetKullaniciByIdAsync(id);
             if (kullanici == null)
             {
-                return NotFound();
+                return this.RecordNotFound(EntityName, "Kullanici", requestedId: id, listLabel: "Kullanıcı listesine dön");
             }
             return View(kullanici);
         }
@@ -99,10 +94,33 @@ namespace Muhasebe.Controllers
         [HttpPost, ActionName("Delete")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            await _kullaniciService.DeleteKullaniciAsync(id);
-            return RedirectToAction("Index");
+            var kullanici = await _kullaniciService.GetKullaniciByIdAsync(id);
+            if (kullanici == null)
+            {
+                return this.RecordNotFound(EntityName, "Kullanici", requestedId: id, listLabel: "Kullanıcı listesine dön");
+            }
 
+            await _kullaniciService.DeleteKullaniciAsync(id);
+            return RedirectToAction(nameof(Index));
+        }
+
+        public async Task<IActionResult> Detail(int id)
+        {
+            var kullanici = await _kullaniciService.GetKullaniciByIdAsync(id);
+            if (kullanici == null)
+            {
+                return this.RecordNotFound(EntityName, "Kullanici", requestedId: id, listLabel: "Kullanıcı listesine dön");
+            }
+            return View(kullanici);
+        }
+
+        private async Task PopulateSelectListsAsync(int? departmanId = null, int? rolId = null)
+        {
+            var departmanlar = await _departmanService.GetAllDepartmanAsync();
+            ViewBag.DepartmanId = new SelectList(departmanlar, "DepartmanId", "DepartmanAdi", departmanId);
+
+            var roller = await _context.Rols.ToListAsync();
+            ViewBag.RolListesi = new SelectList(roller, "RolId", "RolAdi", rolId);
         }
     }
-
 }
